Remove support for duplicate file entries.
BZ: 91074

Improper installation of unsigned code
ID: ANDROID-8219321
Severity: High
Affected versions: Android 2.0 and greater

An inconsistency in the handling of zip files during application
installation may lead to the installation and execution of unsigned
code in a privileged context.

This issue will be publicly disclosed in 90 days. A CTS test will be
included in the next CTS release.

Bug: 8219321
Change-Id: Ibc56bea753917c38e1bb20df48aa45fdff39d364
Reviewed-on: http://android.intel.com:8080/95219
Reviewed-by: Du, Chenyang <chenyang.du@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Tested-by: Shen, JianchunX <jianchunx.shen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>